fix: suppliers menu on every method

+++ src/class/suppliers/suppliers.cpp
@@ -3,16 +3,18 @@
 void suppliers::add_sup()
 {
     cout << "\nAdd Supplier" << endl;
-    cout << "\tSupplier Name: ";
-    cin >> name;
-    cout << "\tPhone number: ";
+    cout << "\tSupplier Name\t: ";
+    cin.ignore();
+    getline(cin, name);
+    cout << "\tPhone number\t: ";
     cin >> phone_num;
-    cout << "\tAddress (in 3 lines, 3rd line is the city): ";
-    cin >> addr_line1;
-    cin >> addr_line2;
-    cin >> addr_city;
-    cout << "\tState: ";
-    cin >> addr_state;
+    cin.ignore();
+    cout << "\tAddress (in 3 lines, 3rd line is the city): \n";
+    getline(cin, addr_line1);
+    getline(cin, addr_line2);
+    getline(cin, addr_city);
+    cout << "\tState\t: ";
+    getline(cin, addr_state);
     
     stmt.str("");
     stmt << "INSERT INTO suppliers(name, phone_num, addr_line1, addr_line2, addr_city, addr_state) VALUES('" << name << "', " << phone_num << ", '"  << addr_line1 << "', '" << addr_line2 << "', '" << addr_city << "', '" << addr_state << "');";
@@ -30,20 +32,61 @@ void suppliers::add_sup()
 void suppliers::remove_supplier()
 {
     cout << "\nRemove Supplier" << endl;
-    cout << "\tSupplier id: ";
+    cout << "\tSupplier id\t: ";
     cin >> id;
+
     stmt.str("");
-    stmt << "DELETE FROM suppliers WHERE id = " << id << ";";
+    stmt << "SELECT name FROM suppliers WHERE id = " << id << ";";
     query = stmt.str();
     q = query.c_str();
-    mysql_query(conn, q);
-    cout << "Supplier Removed." << endl;
+
+    if (mysql_query(conn, q)) {
+        cout << "Query Error: " << mysql_error(conn) << endl;
+        return;
+    }
+
+    res_set = mysql_store_result(conn);
+
+    if ((row = mysql_fetch_row(res_set)) != NULL)
+    {
+        char choice;
+        cout << "\tSupplier name\t: " << row[0] << endl;
+        cout << "\tDo yout want to remove the supplier? [y/n]: ";
+        cin >> choice;
+        if (choice == 121 || choice == 89)
+        {
+            stmt.str("");
+            stmt << "DELETE FROM suppliers WHERE id = " << id << ";";
+            query = stmt.str();
+            q = query.c_str();
+            
+            if (mysql_query(conn, q)) {
+                cout << "Query Error: " << mysql_error(conn) << endl;
+                return;
+            }
+
+            res_set = mysql_store_result(conn);
+
+            if (!(res_set))
+                cout << endl << endl << "Supplier removed." << endl << endl << endl;
+            else
+                cout << endl << endl << "Entry ERROR!" << endl << "Contact Technical Team" << endl << endl << endl;
+        }
+        else
+        {
+            cout << "No change made." << endl;
+        }
+    }
+    else
+    {
+        cout << "No supplier found." << endl;
+    }
 }
 
 void suppliers::search_id()
 {
     cout << "\nSearch Supplier Details by Id" << endl;
-    cout << "\tSupplier id: ";
+    cout << "\tSupplier id\t: ";
     cin >> id;
     stmt.str("");
     stmt << "SELECT * FROM suppliers WHERE id = " << id << ";";
@@ -54,12 +97,11 @@ void suppliers::search_id()
 
     if ((row = mysql_fetch_row(res_set)) != NULL)
     {
-        cout << "\tDetail of Supplier Id: " << row[0] << endl;
-        cout << "\tName: " << row[1] << endl;
-        cout << "\tPhone Number: " << row[2] << endl;
-        cout << "\tAddress: " << row[3] << " " << row[4] << endl;
-        cout << "\tCity: " << row[5] << endl;
-        cout << "\tState: " << row[6] << endl;
+        cout << "\tName\t\t: " << row[1] << endl;
+        cout << "\tPhone Number\t: " << row[2] << endl;
+        cout << "\tAddress\t\t: " << row[3] << ", " << row[4] << endl;
+        cout << "\tCity\t\t: " << row[5] << endl;
+        cout << "\tState\t\t: " << row[6] << endl;
     }
     else
     {
